pyproject.toml
[project]
name = "fcs19-main"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "argon2-cffi>=25.1.0",
    "cryptography>=46.0.7",
    "flask>=3.1.3",
    "flask-cors>=6.0.2",
    "pyotp>=2.9.0",
]
